tools are now limited by shops

+++ backend/tools.js
@@ -4,27 +4,27 @@ const pool = routes.pool;
 // https://kti.com/getTool?id=1
 routes.router.get('/getTool', routes.checkAuthenticated, async (req, res) => {
   let db = await pool.awaitGetConnection();
-  let sql = 'SELECT * FROM singleTools WHERE toolID = ?';
+  let sql = 'SELECT * FROM singleTools WHERE toolID = ? and shopId = ?';
 
-  let data = await db.awaitQuery(sql, req.query.id);
+  let data = await db.awaitQuery(sql, [req.query.id, req.user.shopId]);
   res.json(data[0])
   db.release();
 })
 
 routes.router.get('/getToolTags', routes.checkAuthenticated, async (req, res) => {
   let db = await pool.awaitGetConnection();
-  let sql = 'SELECT * FROM toolTag WHERE toolId = ? INNER JOIN tag ON tag.tagId = toolTag.tagId';
+  let sql = 'SELECT * FROM toolTag WHERE toolId = ? INNER JOIN tag ON tag.tagId = toolTag.tagId INNER JOIN tool on tool.toolId = toolTag.toolId and tool.shopId = ?';
 
-  let data = await db.awaitQuery(sql, req.query.id);
+  let data = await db.awaitQuery(sql, [req.query.id, req.user.shopId]);
   res.json(data)
   db.release();
 })
 
 routes.router.get('/getTaggedTools', routes.checkAuthenticated, async (req, res) => {
   let db = await pool.awaitGetConnection();
-  let sql = 'SELECT * FROM toolTag WHERE toolId = ? INNER JOIN tool ON toolId = toolTag.toolId';
+  let sql = 'SELECT * FROM toolTag WHERE toolId = ? INNER JOIN tool ON tool.toolId = toolTag.toolId and tool.shopId = ?';
 
-  let data = await db.awaitQuery(sql, req.query.id);
+  let data = await db.awaitQuery(sql, [req.query.id, req.user.shopId]);
   res.json(data)
   db.release();
 })
@@ -34,9 +34,9 @@ routes.router.get('/getTaggedTools', routes.checkAuthenticated, async (req, res)
 // https://kti.com/getTools
 routes.router.get('/getTools', routes.checkAuthenticated, async (req, res) => {
   let db = await pool.awaitGetConnection();
-  let sql = 'SELECT * FROM singleTools';
+  let sql = 'SELECT * FROM singleTools where shopId = ?';
 
-  let data = await db.awaitQuery(sql);
+  let data = await db.awaitQuery(sql, req.user.shopId);
   res.json(data)
   db.release();
 })
@@ -45,21 +45,21 @@ routes.router.get("/getToolHistory", routes.checkAdmin, async (req, res) => {
   let db = await pool.awaitGetConnection();
 
   if ("userId" in req.query) {
-    let sql = 'SELECT * FROM toolHistory WHERE userId = ?';
+    let sql = 'SELECT recordId, toolId, userId, timeTaken, timeReturned FROM toolHistory WHERE userId = ? INNER JOIN tool ON tool.toolId = toolHistory.toolId AND tool.shopId = ?';
 
-    let data = await db.awaitQuery(sql, req.query.userId);
+    let data = await db.awaitQuery(sql, [req.query.toolId, req.user.shopId]);
     res.json(data);
   } 
   else if ("toolId" in req.query) {
-    let sql = 'SELECT * FROM toolHistory WHERE toolId = ?';
+    let sql = 'SELECT recordId, toolId, userId, timeTaken, timeReturned FROM toolHistory WHERE toolId = ? INNER JOIN tool ON tool.toolId = toolHistory.toolId AND tool.shopId = ?';
 
-    let data = await db.awaitQuery(sql, req.query.toolId);
+    let data = await db.awaitQuery(sql, [req.query.toolId, req.user.shopId]);
     res.json(data);
   } 
   else {
-    let sql = 'SELECT * FROM toolHistory';
+    let sql = 'SELECT recordId, toolId, userId, timeTaken, timeReturned FROM toolHistory INNER JOIN tool ON tool.toolId = toolHistory.toolId AND tool.shopId = ?';
 
-    let data = await db.awaitQuery(sql); 
+    let data = await db.awaitQuery(sql, req.user.shopId); 
     res.json(data);
   }
 
@@ -81,6 +81,7 @@ routes.router.get("/getMyHistory", routes.checkAuthenticated, async (req, res) =
   
 })
 
+/* deal with later
 const multer  = require('multer')
 const upload = multer({ dest: 'public/images/' })
 routes.router.post('/addToolPicture', upload.single('picture'), async function (req, res, next) {
@@ -99,7 +100,7 @@ routes.router.post('/addToolPicture', upload.single('picture'), async function (
   });
   db.release();
 })
-
+*/ 
 
 // Creates a new tool type
 // https://kti.com/createToolType?name=newTool
@@ -131,6 +132,20 @@ routes.router.post('/createTool', routes.checkAdmin, async (req, res) => {
 routes.router.post('/tagTool', routes.checkAdmin, async (req, res) => {
 
   let db = await pool.awaitGetConnection();
+
+  let test = 'SELECT * FROM tool WHERE toolId = ? AND shopId = ?'
+  let data = await db.awaitQuery(test, [req.query.toolId, req.user.shopId]);
+
+  if (data.length == 0) {
+    res.status(400).send({
+      success: false,
+      msg: "No such tool found"
+    });
+
+    db.release();
+    return;
+  }
+
   let sql = "INSERT INTO toolTag SET ?";
   let tool = {
       toolId: req.body.toolId,
@@ -153,6 +168,19 @@ routes.router.post('/tagTool', routes.checkAdmin, async (req, res) => {
 })
 
 routes.router.delete('/removeTool', routes.checkAdmin, async (req, res) => {
+
+  let test = 'SELECT * FROM tool WHERE toolId = ? AND shopId = ?'
+  let data = await db.awaitQuery(test, [req.query.id, req.user.shopId]);
+
+  if (data.length == 0) {
+    res.status(400).send({
+      success: false,
+      msg: "No such tool found"
+    });
+
+    db.release();
+    return;
+  }
 
   let db = await pool.awaitGetConnection();
   let sql = "DELETE FROM singleTools WHERE toolId  = ?";
@@ -190,6 +218,29 @@ routes.router.get('/getUserTools', routes.checkAuthenticated, async (req, res) =
 
 routes.router.post('/checkoutTool', routes.checkAuthenticated, async (req, res) => {
   let db = await pool.awaitGetConnection();
+
+  let test = 'SELECT * FROM tool WHERE toolId = ? AND shopId = ?'
+  let data = await db.awaitQuery(test, [req.query.id, req.user.shopId]);
+
+  if (data.length == 0) {
+    res.status(400).send({
+      success: false,
+      msg: "No such tool found"
+    });
+
+    db.release();
+    return;
+  }
+  else if (data.takenBy != null) {
+    res.status(403).send({
+      success: false,
+      msg: "Already taken out"
+    });
+
+    db.release();
+    return;
+  }
+
   let sql = 'UPDATE singleTools SET takenBy = ? WHERE toolID = ?';
 
   let success = true;
@@ -227,6 +278,84 @@ routes.router.post('/checkoutTool', routes.checkAuthenticated, async (req, res) 
 
 routes.router.post('/returnTool', routes.checkAuthenticated, async (req, res) => {
   let db = await pool.awaitGetConnection();
+
+  let test = 'SELECT * FROM tool WHERE toolId = ? AND shopId = ?'
+  let data = await db.awaitQuery(test, [req.query.id, req.user.shopId]);
+
+  if (data.length == 0) {
+    res.status(400).send({
+      success: false,
+      msg: "No such tool found"
+    });
+
+    db.release();
+    return;
+  }
+  else if (data.takenBy != req.user.userId) {
+    res.status(403).send({
+      success: false,
+      msg: "Cannot return"
+    });
+
+    db.release();
+    return;
+  }
+
+
+  let sql = "UPDATE singleTools SET takenBy = null WHERE toolID = ?";
+
+  let success = true;
+  let msg = "";
+
+  await db.query(sql, req.query.id, async (error, result) => {
+      if (error) {
+        success = false;
+        msg = "An unexpected error has occured, make sure you passed in all fields correctly";
+      }
+
+      let sql2 = "UPDATE toolHistory SET timeReturned = ? WHERE recordId = (SELECT recordId FROM toolHistory WHERE toolId = ? AND timeReturned IS NULL ORDER BY recordId DESC LIMIT 1)"
+
+      await db.query(sql2, [new Date().toISOString().slice(0, 19).replace('T', ' '), req.query.id], (error, result) => {
+        if (error) {
+          success = false;
+          msg = "An unexpected error has occured, make sure you passed in all fields correctly";
+          console.log(error);
+        }
+      }) 
+
+      res.send({ success: success, msg: msg })
+  });
+
+  db.release();
+})
+
+
+routes.router.post('/forceReturnTool', routes.checkAdmin, async (req, res) => {
+  let db = await pool.awaitGetConnection();
+
+  let test = 'SELECT * FROM tool WHERE toolId = ? AND shopId = ?'
+  let data = await db.awaitQuery(test, [req.query.id, req.user.shopId]);
+
+  if (data.length == 0) {
+    res.status(400).send({
+      success: false,
+      msg: "No such tool found"
+    });
+
+    db.release();
+    return;
+  }
+  else if (data.takenBy == null) {
+    res.status(403).send({
+      success: false,
+      msg: "Not taken out"
+    });
+
+    db.release();
+    return;
+  }
+
+
   let sql = "UPDATE singleTools SET takenBy = null WHERE toolID = ?";
 
   let success = true;
